Added permission checks to mod commands

+++ commands/moderation/purge.js
@@ -1,5 +1,6 @@
 const { SlashCommandBuilder } = require('@discordjs/builders')
 const { simpleEmbed, sleep } = require('../../utilities')
+const { Permissions } = require('discord.js')
 
 module.exports = {
   data: new SlashCommandBuilder()
@@ -11,10 +12,15 @@ module.exports = {
         .setRequired(true)),
   async execute (interaction) {
     let amount = interaction.options.getInteger('amount')
+    amount = amount.toString()
+
+    if (!interaction.user.permissions.has(Permissions.FLAGS.MANAGE_MESSAGES)) {
+      return await interaction.reply(simpleEmbed('You do not have permission to execute this command!', true))
+    }
     if (amount < 1 || amount > 100) {
       return await interaction.reply(simpleEmbed('You can only delete between 1-100 messages!', true))
     }
-    amount = amount.toString()
+
     await interaction.channel.messages.fetch({ limit: amount }).then(messages => {
       interaction.channel.bulkDelete(messages)
     })
